project ID에 따른 화면 전환 구현 및 스캐폴딩 변경
