+++ src/projects/1/index.js
@@ -1,0 +1,9 @@
+function Index() {
+    return (
+        <>
+            Project1 Main
+        </>
+    )
+}
+
+export default Index;
