+++ src/projects/2/index.js
@@ -1,0 +1,9 @@
+function Index() {
+    return (
+        <>
+            Project2 Main
+        </>
+    )
+}
+
+export default Index;
